changed casting
git-svn-id: https://arducopter.googlecode.com/svn/trunk@1515 f9c3cf11-9bcb-44bc-f272-b75c42450872

+++ ArduCopterMega/read_commands.pde
@@ -95,8 +95,8 @@ void parseCommand(char *buffer)
 				break;
 	
 			case 'X':
-				pid_stabilize_roll.imax((int)(value * 100));
-				pid_stabilize_pitch.imax((int)(value * 100));
+				pid_stabilize_roll.imax(value * 100);
+				pid_stabilize_pitch.imax(value * 100);
 				save_EEPROM_PID();
 				break;
